- Изменено пространство имен у общего модуля с DTO + Добавлена зависимость на WebFlex для взаимодействия сервисов через WebClient + в меетоды OperationController добавлен "проброс" параметра "userId" из хедера запросов + частично доработан AccountServiceIntegration, добавлена загрузка параметров из файла конфигурации

+++ operation-service/src/main/java/ru/sberbank/jd/dto/DepositeAccountDto.java
@@ -16,12 +16,11 @@ public class DepositeAccountDto {
     private String depositeAccount;
     private String returnAccount;
     private BigDecimal amount;
-    private BigDecimal depositeRate;
+    private double depositeRate;
 
     private String description;
 
     private LocalDateTime openingDate;
     private LocalDateTime closedDate;
-
 
 }
